# Improve convolve/correlate