This PR reverts the workaround introduced in #3587, and utilizes the new cuFFT plan cache for further speed-up. Both CPU and GPU times are reduced. The test script used below is from https://github.com/cupy/cupy/pull/3560#issue-445154962.

Before this PR:
```
convolve            :    CPU:  307.700 us   +/- 5.083 (min:  302.990 / max:  363.945) us     GPU-0:  313.297 us   +/- 5.175 (min:  308.160 / max:  369.920) us
convolve            :    CPU:  362.793 us   +/- 8.644 (min:  355.626 / max:  506.037) us     GPU-0:  485.262 us   +/-33.614 (min:  461.600 / max:  602.496) us
convolve            :    CPU:  360.504 us   +/- 4.178 (min:  355.401 / max:  396.366) us     GPU-0:  464.572 us   +/- 2.826 (min:  461.248 / max:  493.024) us
convolve            :    CPU:  364.846 us   +/- 5.211 (min:  360.014 / max:  475.614) us     GPU-0:  705.190 us   +/- 4.038 (min:  700.832 / max:  807.104) us
```
After:
```
convolve            :    CPU:  120.850 us   +/- 6.676 (min:  116.539 / max:  188.323) us     GPU-0:  125.311 us   +/- 6.826 (min:  120.928 / max:  193.472) us
convolve            :    CPU:  158.206 us   +/- 2.951 (min:  154.770 / max:  188.331) us     GPU-0:  303.941 us   +/-37.166 (min:  284.064 / max:  394.144) us
convolve            :    CPU:  158.594 us   +/- 4.195 (min:  154.171 / max:  200.894) us     GPU-0:  286.445 us   +/- 1.778 (min:  284.064 / max:  312.064) us
convolve            :    CPU:  160.981 us   +/- 3.870 (min:  157.066 / max:  194.607) us     GPU-0:  524.811 us   +/- 1.750 (min:  522.176 / max:  550.592) us
```